requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dataclasses==0.8
dj-static==0.0.6
Django==3.2.11
django-filter==21.1
django-geojson==3.2.0
django-leaflet==0.28.2
djangorestframework==3.13.1
djangorestframework-gis==0.18
Flask==2.0.2
geos==0.2.3
gunicorn==19.1.1
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.8.5
psycopg2-binary==2.9.3
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
typing-extensions==4.0.1
urllib3==1.25.7
Werkzeug==2.0.2
zipp==3.6.0
